pyproject.toml
[project]
name = "pozar-drevo"
version = "0.1.0"
description = "Posudek drevených nosníků dle ČSN EN 1995-1-1"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "fpdf2>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
